move events logic and rest uris to separate files
